backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
scikit-learn==1.5.2
email-validator==2.2.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
PyPDF2==3.0.1
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.3
